Merge commit '8ddb9e430f6bc728708ae8090ac70fdf3ba7c24d' into develop

Conflicts:
	src/framework/mocha/MochaRuntime.m

+++ src/editor/JSTPluginMover.m
@@ -32,7 +32,7 @@
     NSString *askQuestion = NSLocalizedString(@"Would you like to copy the plugin '%@' into your JSTalk Plug-ins folder?",
                                               @"Would you like to copy the plugin '%@' into your JSTalk Plug-ins folder?");
     
-    int decision = NSRunAlertPanel(NSLocalizedString(@"Install Plug-in?", @"Install Plug-in?"),
+    NSInteger decision = NSRunAlertPanel(NSLocalizedString(@"Install Plug-in?", @"Install Plug-in?"),
                                    [NSString stringWithFormat:askQuestion, thePluginName], @"Install Plug-in", @"Cancel", nil);
